add check for being logged in to all pages

+++ src/main/java/com/example/rentalcarflowmanagerwebapp/controller/ContractController.java
@@ -3,11 +3,14 @@ package com.example.rentalcarflowmanagerwebapp.controller;
 import com.example.rentalcarflowmanagerwebapp.model.Contract;
 import com.example.rentalcarflowmanagerwebapp.service.ContractService;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.HttpMediaTypeException;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
+import javax.servlet.http.HttpSession;
 
 // Tobias
 @Controller
@@ -22,7 +25,12 @@ public class ContractController {
     @PostMapping("create/contract/{employeeID}")
     public String createContract(@RequestParam("customerID") int customerID,
                                  @PathVariable("employeeID") int employeeID,
-                                 RedirectAttributes ra){
+                                 RedirectAttributes ra, HttpSession session){
+
+        // check that its a logged in person accessing the page, redirects to log in page if not
+        if (session.getAttribute("isLoggedIn") == null || !((boolean) session.getAttribute("isLoggedIn"))) {
+            return "index";
+        }
 
         Contract contract = new Contract(-1, customerID, employeeID);
 
@@ -40,10 +48,13 @@ public class ContractController {
         return "redirect:/lease";
     }
 
-
-
     @GetMapping("delete/contract/{contractID}")
-    public String deleteContract(@PathVariable("contractID") int contractID, RedirectAttributes ra){
+    public String deleteContract(@PathVariable("contractID") int contractID, RedirectAttributes ra, HttpSession session){
+
+        // check that its a logged in person accessing the page, redirects to log in page if not
+        if (session.getAttribute("isLoggedIn") == null || !((boolean) session.getAttribute("isLoggedIn"))) {
+            return "index";
+        }
 
         boolean isDeleted = contractService.deleteContract(contractID);
 
@@ -55,15 +66,7 @@ public class ContractController {
             ra.addFlashAttribute("fail", message);
         }
 
-
-
-
-
-
     return "redirect:/lease";
     }
-
-
-
 
 }
